business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetDCSCode.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetDCSCode implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+D ");
        messageString.append(command.getParameter("DCS Code"));
        messageString.append("\n");
        return messageString;
    }
}
